+++ src/addon/fold-html.ts
@@ -329,7 +329,6 @@ export class FoldHTML implements Addon.Addon, Options {
     }
 
     marker = cm.markText(from, to, {
-      clearOnEnter: true,
       replacedWith,
     })